+++ examples/intel/ip/reputation_cymru.go
@@ -8,7 +8,7 @@ import (
 	"os"
 
 	"github.com/pangeacyber/pangea-go/pangea-sdk/pangea"
-	"github.com/pangeacyber/pangea-go/pangea-sdk/service/user_intel"
+	"github.com/pangeacyber/pangea-go/pangea-sdk/service/ip_intel"
 )
 
 func main() {
@@ -17,21 +17,20 @@ func main() {
 		log.Fatal("Unauthorized: No token present")
 	}
 
-	intelcli := user_intel.New(&pangea.Config{
+	intelcli := ip_intel.New(&pangea.Config{
 		Token:  token,
 		Domain: os.Getenv("PANGEA_DOMAIN"),
 	})
 
 	ctx := context.Background()
-	input := &user_intel.UserPasswordBreachedRequest{
-		HashType:   user_intel.HTsha265,
-		HashPrefix: "5baa6",
-		Raw:        true,
-		Verbose:    true,
-		Provider:   "spycloud",
+	input := &ip_intel.IpReputationRequest{
+		Ip:       "93.231.182.110",
+		Raw:      true,
+		Verbose:  true,
+		Provider: "cymru",
 	}
 
-	resp, err := intelcli.PasswordBreached(ctx, input)
+	resp, err := intelcli.Reputation(ctx, input)
 	if err != nil {
 		log.Fatal(err)
 	}