+++ CHANGELOG.md
@@ -5,6 +5,29 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.9.0] - 2023-05-25
+
+# Added
+
+- New algorithm support in Vault Service
+- Algorithm field support in Audit Service
+- Cymru IP Intel provider examples
+- Support full url as domain in config for local use
+
+
+## [1.8.0] - 2023-05-05
+
+### Added
+
+- AuthN support
+- Docs to multiple services
+
+### Fixed
+
+- Redact Score field type
+- Readme example links to SDK examples page
+
+
 ## [1.7.0] - 2023-04-10
 
 ### Added
@@ -17,7 +40,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.6.0] - 2023-03-27
 
-###
+### Added
 
 - Algorithm support in Vault Service
 
@@ -102,7 +125,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Redact client
 
 
-[unreleased]: https://github.com/pangeacyber/pangea-go/compare/pangea-sdk/v1.7.0...main
+[unreleased]: https://github.com/pangeacyber/pangea-go/compare/pangea-sdk/v1.9.0...main
+[1.9.0]: https://github.com/pangeacyber/pangea-go/compare/pangea-sdk/v1.8.0...v1.9.0
+[1.8.0]: https://github.com/pangeacyber/pangea-go/compare/pangea-sdk/v1.7.0...v1.8.0
 [1.7.0]: https://github.com/pangeacyber/pangea-go/compare/pangea-sdk/v1.6.0...v1.7.0
 [1.6.0]: https://github.com/pangeacyber/pangea-go/compare/pangea-sdk/v1.5.0...v1.6.0
 [1.5.0]: https://github.com/pangeacyber/pangea-go/compare/pangea-sdk/v1.4.0...v1.5.0